Solved a problem with AbstractTransportListener#startListeningForService.
Problem description: When a runtime exception is thrown by the implementation of the startListeningForService (either explicitly or e.g. a NPE), it is not caught by AbstractTransportListener, but in AxisConfiguration#notifyObservers. The exception then gets logged at DEBUG level without a full stack trace. Also the service will not be disabled for the transport and not be marked as faulty.

In addition, the transports built on AbstractTransportListener behave quite differently when a problem occurs in startListeningForService:
* JMSListener marks the service as faulty and disables it for the transport.
* AbstractPollingTransportListener and AbstractDatagramTransportListener only disable the service for the transport, but will not mark it as faulty.
* FIXTransportListener only logs the error.

Changes:
* Intercept exceptions thrown by startListeningForService (including runtime exceptions) in AbstractTransportListener and take appropriate action, i.e. disable the service for the transport and mark it as faulty.
* Allow startListeningForService to throw AxisFaults, so that implementations can simply throw this exception and don't need to care themselves about disabling the service or marking it as faulty.


git-svn-id: https://svn.apache.org/repos/asf/webservices/commons/trunk/modules/transport@712501 13f79535-47bb-0310-9956-ffa450edef68