Moved all invocations of the maven-javadoc-plugin from the <reporting/> tag to the <build/> tag (phase "site") and make sure that every configuration is bound to a single execution.
Rationale: When a configuration for maven-javadoc-plugin is specified in the <reporting/> tag (as was the case in the testkit module), Maven 2.2 (but not 3.0) applies that configuration also to the maven-javadoc-plugin execution defined by the org.apache:apache super-POM (and which is triggered by the apache-release profile). The change described above will avoid this confusion and seems to work with all Maven versions (2.2.x and 3.0.x), goals (clean install or clean site) and profiles (apache-release or no profile), although I didn't test all combinations.

Side effect: mvn site:site will no longer produce the Javadoc. Instead, mvn site must be used (which should be used anyway, at least to generate the site for publication).


git-svn-id: https://svn.apache.org/repos/asf/axis/axis2/java/transports/trunk@1127257 13f79535-47bb-0310-9956-ffa450edef68